+++ crates/lib/src/games/unreal2.rs
@@ -1,0 +1,10 @@
+//! Unreal2 game query modules
+
+use crate::protocols::unreal2::game_query_mod;
+
+game_query_mod!(darkesthour, "Darkest Hour: Europe '44-'45 (2008)", 7758);
+game_query_mod!(devastation, "Devastation (2003)", 7778);
+game_query_mod!(killingfloor, "Killing Floor", 7708);
+game_query_mod!(redorchestra, "Red Orchestra", 7759);
+game_query_mod!(ut2003, "Unreal Tournament 2003", 7758);
+game_query_mod!(ut2004, "Unreal Tournament 2004", 7778);