+++ crates/lib/src/protocols/minecraft/mod.rs
@@ -4,5 +4,4 @@ pub mod protocol;
 pub mod types;
 
 pub use protocol::*;
-pub use protocol::*;
 pub use types::*;